+++ main.h
@@ -5,7 +5,7 @@ int _printf(const char *format, ...);
 void print_char(va_list ap, int *falcon);
 void print_s(va_list ap, int *falcon);
 void print_per(int *falcon);
-int sou(int d_d, int falcon);
+int sou(long int d_d, int falcon);
 int print_d(va_list ap, int falcon);
 void print_undef(char s, int *falcon);
 #endif
